requirements.txt
Flask
flask-cors
torch
torchvision
Pillow